file names updated to reflect new class names, added in new unit test file for GetCartItem and first test to this
